Refactor cometbft service and move handlers to separate files
Now we have separate go files for each abci call we want to handle
and this cleans up the code nicely and makes it easier to read and
traverse the code.
