+++ consensus/cometbft/service/abci.go
@@ -25,20 +25,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"time"
 
-	"cosmossdk.io/store/rootmulti"
-	ctypes "github.com/berachain/beacon-kit/consensus-types/types"
-	servercmtlog "github.com/berachain/beacon-kit/consensus/cometbft/service/log"
-	"github.com/berachain/beacon-kit/consensus/types"
-	errorsmod "github.com/berachain/beacon-kit/errors"
-	"github.com/berachain/beacon-kit/primitives/encoding/json"
-	"github.com/berachain/beacon-kit/primitives/math"
 	cmtabci "github.com/cometbft/cometbft/abci/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	abci "github.com/cometbft/cometbft/api/cometbft/abci/v1"
 	sdkversion "github.com/cosmos/cosmos-sdk/version"
-	"github.com/sourcegraph/conc/iter"
 )
 
 var (
@@ -53,30 +43,16 @@ func (s *Service[LoggerT]) InitChain(
 	return s.initChain(ctx, req)
 }
 
-// InitChainer initializes the chain.
-func (s *Service[LoggerT]) initChainer(
-	ctx sdk.Context,
-	appStateBytes []byte,
-) ([]cmtabci.ValidatorUpdate, error) {
-	var genesisState map[string]json.RawMessage
-	if err := json.Unmarshal(appStateBytes, &genesisState); err != nil {
-		return nil, err
-	}
-
-	data := []byte(genesisState["beacon"])
-	valUpdates, err := s.Blockchain.ProcessGenesisData(ctx, data)
-	if err != nil {
-		return nil, err
-	}
-
-	return iter.MapErr(
-		valUpdates,
-		convertValidatorUpdate[cmtabci.ValidatorUpdate],
-	)
+// PrepareProposal implements the PrepareProposal ABCI method and returns a
+// ResponsePrepareProposal object to the client.
+func (s *Service[LoggerT]) PrepareProposal(
+	ctx context.Context,
+	req *cmtabci.PrepareProposalRequest,
+) (*cmtabci.PrepareProposalResponse, error) {
+	return s.prepareProposal(ctx, req)
 }
 
-func (s *Service[LoggerT]) Info(
-	context.Context,
+func (s *Service[LoggerT]) Info(context.Context,
 	*cmtabci.InfoRequest,
 ) (*cmtabci.InfoResponse, error) {
 	lastCommitID := s.sm.CommitMultiStore().LastCommitID()
@@ -98,233 +74,20 @@ func (s *Service[LoggerT]) Info(
 	}, nil
 }
 
-// PrepareProposal implements the PrepareProposal ABCI method and returns a
-// ResponsePrepareProposal object to the client.
-func (s *Service[LoggerT]) PrepareProposal(
-	_ context.Context,
-	req *cmtabci.PrepareProposalRequest,
-) (*cmtabci.PrepareProposalResponse, error) {
-	startTime := time.Now()
-	defer s.telemetrySink.MeasureSince(
-		"beacon_kit.runtime.prepare_proposal_duration", startTime)
-
-	// CometBFT must never call PrepareProposal with a height of 0.
-	if req.Height < 1 {
-		return nil, fmt.Errorf(
-			"prepareProposal at height %v: %w",
-			req.Height,
-			errInvalidHeight,
-		)
-	}
-
-	// Always reset state given that PrepareProposal can timeout
-	// and be called again in a subsequent round.
-	s.prepareProposalState = s.resetState()
-	s.prepareProposalState.SetContext(
-		s.getContextForProposal(
-			s.prepareProposalState.Context(),
-			req.Height,
-		),
-	)
-
-	var slotData = types.NewSlotData[
-		ctypes.AttestationData,
-		ctypes.SlashingInfo,
-	](
-		math.Slot(req.GetHeight()),
-		nil,
-		nil,
-		req.GetProposerAddress(),
-		req.GetTime(),
-	)
-
-	// TODO: can we use the context from the request now that async is gone?
-	//nolint:contextcheck // see todo above
-	blkBz, sidecarsBz, err := s.BlockBuilder.BuildBlockAndSidecars(
-		s.prepareProposalState.Context(),
-		*slotData,
-	)
-	if err != nil {
-		s.logger.Error(
-			"failed to prepare proposal",
-			"height",
-			req.Height,
-			"time",
-			req.Time,
-			"err",
-			err,
-		)
-		return &cmtabci.PrepareProposalResponse{Txs: req.Txs}, nil
-	}
-
-	return &cmtabci.PrepareProposalResponse{
-		Txs: [][]byte{blkBz, sidecarsBz},
-	}, nil
-}
-
 // ProcessProposal implements the ProcessProposal ABCI method and returns a
 // ResponseProcessProposal object to the client.
 func (s *Service[LoggerT]) ProcessProposal(
-	_ context.Context,
+	ctx context.Context,
 	req *cmtabci.ProcessProposalRequest,
 ) (*cmtabci.ProcessProposalResponse, error) {
-	// CometBFT must never call ProcessProposal with a height of 0.
-	if req.Height < 1 {
-		return nil, fmt.Errorf(
-			"processProposal at height %v: %w",
-			req.Height,
-			errInvalidHeight,
-		)
-	}
-
-	// Since the application can get access to FinalizeBlock state and write to
-	// it, we must be sure to reset it in case ProcessProposal timeouts and is
-	// called
-	// again in a subsequent round. However, we only want to do this after we've
-	// processed the first block, as we want to avoid overwriting the
-	// finalizeState
-	// after state changes during InitChain.
-	s.processProposalState = s.resetState()
-	if req.Height > s.initialHeight {
-		s.finalizeBlockState = s.resetState()
-	}
-
-	s.processProposalState.SetContext(
-		s.getContextForProposal(
-			s.processProposalState.Context(),
-			req.Height,
-		),
-	)
-
-	resp, err := s.Middleware.ProcessProposal(
-		s.processProposalState.Context(),
-		req,
-	)
-	if err != nil {
-		s.logger.Error(
-			"failed to process proposal",
-			"height",
-			req.Height,
-			"time",
-			req.Time,
-			"hash",
-			fmt.Sprintf("%X", req.Hash),
-			"err",
-			err,
-		)
-		return &cmtabci.ProcessProposalResponse{
-			Status: cmtabci.PROCESS_PROPOSAL_STATUS_REJECT,
-		}, nil
-	}
-
-	return resp, nil
-}
-
-func (s *Service[LoggerT]) internalFinalizeBlock(
-	req *cmtabci.FinalizeBlockRequest,
-) (*cmtabci.FinalizeBlockResponse, error) {
-	if err := s.validateFinalizeBlockHeight(req); err != nil {
-		return nil, err
-	}
-
-	// finalizeBlockState should be set on InitChain or ProcessProposal. If it
-	// is nil, it means we are replaying this block and we need to set the state
-	// here given that during block replay ProcessProposal is not executed by
-	// CometBFT.
-	if s.finalizeBlockState == nil {
-		s.finalizeBlockState = s.resetState()
-	}
-
-	// Iterate over all raw transactions in the proposal and attempt to execute
-	// them, gathering the execution results.
-	//
-	// NOTE: Not all raw transactions may adhere to the sdk.Tx interface, e.g.
-	// vote extensions, so skip those.
-	txResults := make([]*cmtabci.ExecTxResult, len(req.Txs))
-	for i := range req.Txs {
-		//nolint:mnd // its okay for now.
-		txResults[i] = &cmtabci.ExecTxResult{
-			Codespace: "sdk",
-			Code:      2,
-			Log:       "skip decoding",
-			GasWanted: 0,
-			GasUsed:   0,
-		}
-	}
-
-	finalizeBlock, err := s.Middleware.FinalizeBlock(
-		s.finalizeBlockState.Context(),
-		req,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	valUpdates, err := iter.MapErr(
-		finalizeBlock,
-		convertValidatorUpdate[cmtabci.ValidatorUpdate],
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return &cmtabci.FinalizeBlockResponse{
-		TxResults:             txResults,
-		ValidatorUpdates:      valUpdates,
-		ConsensusParamUpdates: s.paramStore.Get(),
-	}, nil
-}
-
-func (s *Service[_]) validateFinalizeBlockHeight(
-	req *cmtabci.FinalizeBlockRequest,
-) error {
-	if req.Height < 1 {
-		return fmt.Errorf(
-			"finalizeBlock at height %v: %w",
-			req.Height,
-			errInvalidHeight,
-		)
-	}
-
-	lastBlockHeight := s.LastBlockHeight()
-
-	// expectedHeight holds the expected height to validate
-	var expectedHeight int64
-	if lastBlockHeight == 0 && s.initialHeight > 1 {
-		// In this case, we're validating the first block of the chain, i.e no
-		// previous commit. The height we're expecting is the initial height.
-		expectedHeight = s.initialHeight
-	} else {
-		// This case can mean two things:
-		//
-		// - Either there was already a previous commit in the store, in which
-		// case we increment the version from there.
-		// - Or there was no previous commit, in which case we start at version
-		// 1.
-		expectedHeight = lastBlockHeight + 1
-	}
-
-	if req.Height != expectedHeight {
-		return fmt.Errorf(
-			"invalid height: %d; expected: %d",
-			req.Height,
-			expectedHeight,
-		)
-	}
-
-	return nil
+	return s.processProposal(ctx, req)
 }
 
 func (s *Service[_]) FinalizeBlock(
-	_ context.Context,
+	ctx context.Context,
 	req *cmtabci.FinalizeBlockRequest,
 ) (*cmtabci.FinalizeBlockResponse, error) {
-	res, err := s.internalFinalizeBlock(req)
-	if res != nil {
-		res.AppHash = s.workingHash()
-	}
-
-	return res, err
+	return s.finalizeBlock(ctx, req)
 }
 
 // Commit implements the ABCI interface. It will commit all state that exists in
@@ -335,210 +98,67 @@ func (s *Service[_]) FinalizeBlock(
 // against that height and gracefully halt if it matches the latest committed
 // height.
 func (s *Service[LoggerT]) Commit(
-	context.Context, *cmtabci.CommitRequest,
+	ctx context.Context, req *cmtabci.CommitRequest,
 ) (*cmtabci.CommitResponse, error) {
-	if s.finalizeBlockState == nil {
-		// This is unexpected since CometBFT should call Commit only
-		// after FinalizeBlock has been called. Panic appeases nilaway.
-		panic(fmt.Errorf("commit: %w", errNilFinalizeBlockState))
-	}
-	header := s.finalizeBlockState.Context().BlockHeader()
-	retainHeight := s.GetBlockRetentionHeight(header.Height)
-
-	rms, ok := s.sm.CommitMultiStore().(*rootmulti.Store)
-	if ok {
-		rms.SetCommitHeader(header)
-	}
-	s.sm.CommitMultiStore().Commit()
-
-	s.finalizeBlockState = nil
-
-	return &cmtabci.CommitResponse{
-		RetainHeight: retainHeight,
-	}, nil
+	return s.commit(ctx, req)
 }
 
-// workingHash gets the apphash that will be finalized in commit.
-// These writes will be persisted to the root multi-store
-// (s.sm.CommitMultiStore()) and flushed
-// to disk in the Commit phase. This means when the ABCI client requests
-// Commit(), the application state transitions will be flushed to disk and as a
-// result, but we already have
-// an application Merkle root.
-func (s *Service[LoggerT]) workingHash() []byte {
-	// Write the FinalizeBlock state into branched storage and commit the
-	// MultiStore. The write to the FinalizeBlock state writes all state
-	// transitions to the root MultiStore (s.sm.CommitMultiStore())
-	// so when Commit() is called it persists those values.
-	if s.finalizeBlockState == nil {
-		// this is unexpected since workingHash is called only after
-		// internalFinalizeBlock. Panic appeases nilaway.
-		panic(fmt.Errorf("workingHash: %w", errNilFinalizeBlockState))
-	}
-	s.finalizeBlockState.ms.Write()
+//
+// NOOP methods
+//
 
-	// Get the hash of all writes in order to return the apphash to the comet in
-	// finalizeBlock.
-	commitHash := s.sm.CommitMultiStore().WorkingHash()
-	s.logger.Debug(
-		"hash of all writes",
-		"workingHash",
-		fmt.Sprintf("%X", commitHash),
-	)
-
-	return commitHash
+func (Service[_]) Query(
+	context.Context,
+	*abci.QueryRequest,
+) (*abci.QueryResponse, error) {
+	return &abci.QueryResponse{}, nil
 }
 
-// getContextForProposal returns the correct Context for PrepareProposal and
-// ProcessProposal. We use finalizeBlockState on the first block to be able to
-// access any state changes made in InitChain.
-func (s *Service[LoggerT]) getContextForProposal(
-	ctx sdk.Context,
-	height int64,
-) sdk.Context {
-	if height != s.initialHeight {
-		return ctx
-	}
-
-	if s.finalizeBlockState == nil {
-		// this is unexpected since cometBFT won't call PrepareProposal
-		// on initialHeight. Panic appeases nilaway.
-		panic(fmt.Errorf("getContextForProposal: %w", errNilFinalizeBlockState))
-	}
-	ctx, _ = s.finalizeBlockState.Context().CacheContext()
-	return ctx
+func (Service[_]) ListSnapshots(
+	context.Context,
+	*abci.ListSnapshotsRequest,
+) (*abci.ListSnapshotsResponse, error) {
+	return &abci.ListSnapshotsResponse{}, nil
 }
 
-// CreateQueryContext creates a new sdk.Context for a query, taking as args
-// the block height and whether the query needs a proof or not.
-func (s *Service[LoggerT]) CreateQueryContext(
-	height int64,
-	prove bool,
-) (sdk.Context, error) {
-	// use custom query multi-store if provided
-	lastBlockHeight := s.sm.CommitMultiStore().LatestVersion()
-	if lastBlockHeight == 0 {
-		return sdk.Context{}, errorsmod.Wrapf(
-			sdkerrors.ErrInvalidHeight,
-			"%s is not ready; please wait for first block",
-			appName,
-		)
-	}
-
-	if height > lastBlockHeight {
-		return sdk.Context{},
-			errorsmod.Wrap(
-				sdkerrors.ErrInvalidHeight,
-				"cannot query with height in the future; please provide a valid height",
-			)
-	}
-
-	// when a client did not provide a query height, manually inject the latest
-	if height == 0 {
-		height = lastBlockHeight
-	}
-
-	if height <= 1 && prove {
-		return sdk.Context{},
-			errorsmod.Wrap(
-				sdkerrors.ErrInvalidRequest,
-				"cannot query with proof when height <= 1; please provide a valid height",
-			)
-	}
-
-	cacheMS, err := s.sm.CommitMultiStore().CacheMultiStoreWithVersion(height)
-	if err != nil {
-		return sdk.Context{},
-			errorsmod.Wrapf(
-				sdkerrors.ErrNotFound,
-				"failed to load state at height %d; %s (latest height: %d)",
-				height,
-				err,
-				lastBlockHeight,
-			)
-	}
-
-	return sdk.NewContext(
-		cacheMS,
-		true,
-		servercmtlog.WrapSDKLogger(s.logger),
-	), nil
+func (Service[_]) LoadSnapshotChunk(
+	context.Context,
+	*abci.LoadSnapshotChunkRequest,
+) (*abci.LoadSnapshotChunkResponse, error) {
+	return &abci.LoadSnapshotChunkResponse{}, nil
 }
 
-// GetBlockRetentionHeight returns the height for which all blocks below this
-// height
-// are pruned from CometBFT. Given a commitment height and a non-zero local
-// minRetainBlocks configuration, the retentionHeight is the smallest height
-// that
-// satisfies:
-//
-// - Unbonding (safety threshold) time: The block interval in which validators
-// can be economically punished for misbehavior. Blocks in this interval must be
-// auditable e.g. by the light client.
-//
-// - Logical store snapshot interval: The block interval at which the underlying
-// logical store database is persisted to disk, e.g. every 10000 heights. Blocks
-// since the last IAVL snapshot must be available for replay on application
-// restart.
-//
-// - State sync snapshots: Blocks since the oldest available snapshot must be
-// available for state sync nodes to catch up (oldest because a node may be
-// restoring an old snapshot while a new snapshot was taken).
-//
-// - Local (minRetainBlocks) config: Archive nodes may want to retain more or
-// all blocks, e.g. via a local config option min-retain-blocks. There may also
-// be a need to vary retention for other nodes, e.g. sentry nodes which do not
-// need historical blocks.
-func (s *Service[_]) GetBlockRetentionHeight(commitHeight int64) int64 {
-	// pruning is disabled if minRetainBlocks is zero
-	if s.minRetainBlocks == 0 {
-		return 0
-	}
+func (Service[_]) OfferSnapshot(
+	context.Context,
+	*abci.OfferSnapshotRequest,
+) (*abci.OfferSnapshotResponse, error) {
+	return &abci.OfferSnapshotResponse{}, nil
+}
 
-	minNonZero := func(x, y int64) int64 {
-		switch {
-		case x == 0:
-			return y
+func (Service[_]) ApplySnapshotChunk(
+	context.Context,
+	*abci.ApplySnapshotChunkRequest,
+) (*abci.ApplySnapshotChunkResponse, error) {
+	return &abci.ApplySnapshotChunkResponse{}, nil
+}
 
-		case y == 0:
-			return x
+func (Service[_]) ExtendVote(
+	context.Context,
+	*abci.ExtendVoteRequest,
+) (*abci.ExtendVoteResponse, error) {
+	return &abci.ExtendVoteResponse{}, nil
+}
 
-		case x < y:
-			return x
+func (Service[_]) VerifyVoteExtension(
+	context.Context,
+	*abci.VerifyVoteExtensionRequest,
+) (*abci.VerifyVoteExtensionResponse, error) {
+	return &abci.VerifyVoteExtensionResponse{}, nil
+}
 
-		default:
-			return y
-		}
-	}
-
-	// Define retentionHeight as the minimum value that satisfies all non-zero
-	// constraints. All blocks below (commitHeight-retentionHeight) are pruned
-	// from CometBFT.
-	var retentionHeight int64
-
-	// Define the number of blocks needed to protect against misbehaving
-	// validators
-	// which allows light clients to operate safely. Note, we piggy back of the
-	// evidence parameters instead of computing an estimated number of blocks
-	// based
-	// on the unbonding period and block commitment time as the two should be
-	// equivalent.
-	if s.finalizeBlockState == nil {
-		return 0
-	}
-	cp := s.paramStore.Get()
-	if cp.Evidence != nil && cp.Evidence.MaxAgeNumBlocks > 0 {
-		retentionHeight = commitHeight - cp.Evidence.MaxAgeNumBlocks
-	}
-
-	//#nosec:G701 // bet.
-	v := commitHeight - int64(s.minRetainBlocks)
-	retentionHeight = minNonZero(retentionHeight, v)
-
-	if retentionHeight <= 0 {
-		// prune nothing in the case of a non-positive height
-		return 0
-	}
-
-	return retentionHeight
+func (*Service[_]) CheckTx(
+	context.Context,
+	*abci.CheckTxRequest,
+) (*abci.CheckTxResponse, error) {
+	return &abci.CheckTxResponse{}, nil
 }
