+++ consensus/cometbft/service/commit.go
@@ -1,0 +1,132 @@
+// SPDX-License-Identifier: BUSL-1.1
+//
+// Copyright (C) 2024, Berachain Foundation. All rights reserved.
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file of this repository and at www.mariadb.com/bsl11.
+//
+// ANY USE OF THE LICENSED WORK IN VIOLATION OF THIS LICENSE WILL AUTOMATICALLY
+// TERMINATE YOUR RIGHTS UNDER THIS LICENSE FOR THE CURRENT AND ALL OTHER
+// VERSIONS OF THE LICENSED WORK.
+//
+// THIS LICENSE DOES NOT GRANT YOU ANY RIGHT IN ANY TRADEMARK OR LOGO OF
+// LICENSOR OR ITS AFFILIATES (PROVIDED THAT YOU MAY USE A TRADEMARK OR LOGO OF
+// LICENSOR AS EXPRESSLY REQUIRED BY THIS LICENSE).
+//
+// TO THE EXTENT PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON
+// AN “AS IS” BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS,
+// EXPRESS OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
+// TITLE.
+//
+
+package cometbft
+
+import (
+	"context"
+	"fmt"
+
+	"cosmossdk.io/store/rootmulti"
+	cmtabci "github.com/cometbft/cometbft/abci/types"
+)
+
+func (s *Service[LoggerT]) commit(
+	context.Context, *cmtabci.CommitRequest,
+) (*cmtabci.CommitResponse, error) {
+	if s.finalizeBlockState == nil {
+		// This is unexpected since CometBFT should call Commit only
+		// after FinalizeBlock has been called. Panic appeases nilaway.
+		panic(fmt.Errorf("commit: %w", errNilFinalizeBlockState))
+	}
+	header := s.finalizeBlockState.Context().BlockHeader()
+	retainHeight := s.GetBlockRetentionHeight(header.Height)
+
+	rms, ok := s.sm.CommitMultiStore().(*rootmulti.Store)
+	if ok {
+		rms.SetCommitHeader(header)
+	}
+	s.sm.CommitMultiStore().Commit()
+
+	s.finalizeBlockState = nil
+
+	return &cmtabci.CommitResponse{
+		RetainHeight: retainHeight,
+	}, nil
+}
+
+// GetBlockRetentionHeight returns the height for which all blocks below this
+// height
+// are pruned from CometBFT. Given a commitment height and a non-zero local
+// minRetainBlocks configuration, the retentionHeight is the smallest height
+// that
+// satisfies:
+//
+// - Unbonding (safety threshold) time: The block interval in which validators
+// can be economically punished for misbehavior. Blocks in this interval must be
+// auditable e.g. by the light client.
+//
+// - Logical store snapshot interval: The block interval at which the underlying
+// logical store database is persisted to disk, e.g. every 10000 heights. Blocks
+// since the last IAVL snapshot must be available for replay on application
+// restart.
+//
+// - State sync snapshots: Blocks since the oldest available snapshot must be
+// available for state sync nodes to catch up (oldest because a node may be
+// restoring an old snapshot while a new snapshot was taken).
+//
+// - Local (minRetainBlocks) config: Archive nodes may want to retain more or
+// all blocks, e.g. via a local config option min-retain-blocks. There may also
+// be a need to vary retention for other nodes, e.g. sentry nodes which do not
+// need historical blocks.
+func (s *Service[_]) GetBlockRetentionHeight(commitHeight int64) int64 {
+	// pruning is disabled if minRetainBlocks is zero
+	if s.minRetainBlocks == 0 {
+		return 0
+	}
+
+	minNonZero := func(x, y int64) int64 {
+		switch {
+		case x == 0:
+			return y
+
+		case y == 0:
+			return x
+
+		case x < y:
+			return x
+
+		default:
+			return y
+		}
+	}
+
+	// Define retentionHeight as the minimum value that satisfies all non-zero
+	// constraints. All blocks below (commitHeight-retentionHeight) are pruned
+	// from CometBFT.
+	var retentionHeight int64
+
+	// Define the number of blocks needed to protect against misbehaving
+	// validators
+	// which allows light clients to operate safely. Note, we piggy back of the
+	// evidence parameters instead of computing an estimated number of blocks
+	// based
+	// on the unbonding period and block commitment time as the two should be
+	// equivalent.
+	if s.finalizeBlockState == nil {
+		return 0
+	}
+	cp := s.paramStore.Get()
+	if cp.Evidence != nil && cp.Evidence.MaxAgeNumBlocks > 0 {
+		retentionHeight = commitHeight - cp.Evidence.MaxAgeNumBlocks
+	}
+
+	//#nosec:G701 // bet.
+	v := commitHeight - int64(s.minRetainBlocks)
+	retentionHeight = minNonZero(retentionHeight, v)
+
+	if retentionHeight <= 0 {
+		// prune nothing in the case of a non-positive height
+		return 0
+	}
+
+	return retentionHeight
+}
