--- consensus/cometbft/service/noops.go
@@ -1,83 +1,0 @@
-// SPDX-License-Identifier: BUSL-1.1
-//
-// Copyright (C) 2024, Berachain Foundation. All rights reserved.
-// Use of this software is governed by the Business Source License included
-// in the LICENSE file of this repository and at www.mariadb.com/bsl11.
-//
-// ANY USE OF THE LICENSED WORK IN VIOLATION OF THIS LICENSE WILL AUTOMATICALLY
-// TERMINATE YOUR RIGHTS UNDER THIS LICENSE FOR THE CURRENT AND ALL OTHER
-// VERSIONS OF THE LICENSED WORK.
-//
-// THIS LICENSE DOES NOT GRANT YOU ANY RIGHT IN ANY TRADEMARK OR LOGO OF
-// LICENSOR OR ITS AFFILIATES (PROVIDED THAT YOU MAY USE A TRADEMARK OR LOGO OF
-// LICENSOR AS EXPRESSLY REQUIRED BY THIS LICENSE).
-//
-// TO THE EXTENT PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON
-// AN “AS IS” BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS,
-// EXPRESS OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
-// TITLE.
-
-package cometbft
-
-import (
-	"context"
-
-	abci "github.com/cometbft/cometbft/api/cometbft/abci/v1"
-)
-
-func (Service[_]) Query(
-	context.Context,
-	*abci.QueryRequest,
-) (*abci.QueryResponse, error) {
-	return &abci.QueryResponse{}, nil
-}
-
-func (Service[_]) ListSnapshots(
-	context.Context,
-	*abci.ListSnapshotsRequest,
-) (*abci.ListSnapshotsResponse, error) {
-	return &abci.ListSnapshotsResponse{}, nil
-}
-
-func (Service[_]) LoadSnapshotChunk(
-	context.Context,
-	*abci.LoadSnapshotChunkRequest,
-) (*abci.LoadSnapshotChunkResponse, error) {
-	return &abci.LoadSnapshotChunkResponse{}, nil
-}
-
-func (Service[_]) OfferSnapshot(
-	context.Context,
-	*abci.OfferSnapshotRequest,
-) (*abci.OfferSnapshotResponse, error) {
-	return &abci.OfferSnapshotResponse{}, nil
-}
-
-func (Service[_]) ApplySnapshotChunk(
-	context.Context,
-	*abci.ApplySnapshotChunkRequest,
-) (*abci.ApplySnapshotChunkResponse, error) {
-	return &abci.ApplySnapshotChunkResponse{}, nil
-}
-
-func (Service[_]) ExtendVote(
-	context.Context,
-	*abci.ExtendVoteRequest,
-) (*abci.ExtendVoteResponse, error) {
-	return &abci.ExtendVoteResponse{}, nil
-}
-
-func (Service[_]) VerifyVoteExtension(
-	context.Context,
-	*abci.VerifyVoteExtensionRequest,
-) (*abci.VerifyVoteExtensionResponse, error) {
-	return &abci.VerifyVoteExtensionResponse{}, nil
-}
-
-func (*Service[_]) CheckTx(
-	context.Context,
-	*abci.CheckTxRequest,
-) (*abci.CheckTxResponse, error) {
-	return &abci.CheckTxResponse{}, nil
-}
